% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{wkstrategyci_command}
\alias{wkstrategyci_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkstrategyci_command(session, cmd_json)
}
\arguments{
\item{session}{A session handle from [wkstrategyci_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
