% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{wkstrategyci_version}
\alias{wkstrategyci_version}
\title{The wickra-strategy-ci library version.}
\usage{
wkstrategyci_version()
}
\value{
A version string.
}
\description{
The wickra-strategy-ci library version.
}
